Rename Home -> HomePage, and BoardEditor -> BoardEditorPage
