--- ui/src/BoardEditor/index.ts
@@ -1,1 +1,0 @@
-export { default as default } from './BoardEditor.svelte';
