+++ ui/src/BoardEditorPage/EmojiPicker.svelte
