+++ ui/src/BoardEditorPage/IntegerInput.svelte
