+++ ui/src/BoardEditorPage/PieceTypeCard.svelte
