--- ui/src/Home/index.ts
@@ -1,1 +1,0 @@
-export { default as default } from './Home.svelte';
