+++ ui/src/HomePage/StartGameDialog.svelte
