+++ ui/src/HomePage/index.ts
@@ -1,0 +1,1 @@
+export { default as default } from './HomePage.svelte';
